created roles

+++ OriginGamer/app/Http/Controllers/CategoryController.php
@@ -8,20 +8,28 @@ use App\Models\Category;
 
 class CategoryController extends Controller
 {
-    // public function __construct()
-    // {
-    //     $this->middleware('auth:api', ['except' => ['index', 'show']]);
-    // }
+
     public function index()
     {
+        if(auth()->user()->hasPermissionTo('view category') == false){
+            return response()->json(['message' => 'You are not authorized to view categories'], 403);
+        }
+        else 
+        {
         $categories = Category::orderBy('id')->get();
         return response()->json([
             'status' => 'success',
             'categories' => $categories,
         ]);
     }
+    }
     public function store(StorecategoryRequest $request)
     {
+        if(auth()->user()->hasPermissionTo('create category') == false){
+            return response()->json(['message' => 'You are not authorized to create categories'], 403);
+        }
+        else 
+        {
         $category = Category::create($request->all());
 
         return response()->json([
@@ -30,18 +38,30 @@ class CategoryController extends Controller
             'category' => $category
         ], 201);
     }
+    }
 
     public function show(Category $category)
     {
+        if(auth()->user()->hasPermissionTo('view category') == false){
+            return response()->json(['message' => 'You are not authorized to view categories'], 403);
+        }
+        else 
+        {
         $category->find($category->id);
         if (!$category) {
             return response()->json(['message' => 'Category not found'], 404);
         }
         return response()->json($category, 200);
     }
+    }
 
     public function update(StorecategoryRequest $request, Category $category)
     {
+        if(auth()->user()->hasPermissionTo('update category') == false){
+            return response()->json(['message' => 'You are not authorized to edit categories'], 403);
+        }
+        else 
+        {
         $category->update($request->all());
 
         if (!$category) {
@@ -54,9 +74,15 @@ class CategoryController extends Controller
             'category' => $category
         ], 200);
     }
+    }
 
     public function destroy(Category $category)
     {
+        if(auth()->user()->hasPermissionTo('delete category') == false){
+            return response()->json(['message' => 'You are not authorized to delete categories'], 403);
+        }
+        else 
+        {
         $category->delete();
 
         if (!$category) {
@@ -70,4 +96,5 @@ class CategoryController extends Controller
             'message' => 'Category deleted successfully'
         ], 200);
     }
+}
 }
